Create distinct Louisiana CDCC refundable and non-refundable variables Fixes #4497
