+++ changelog_entry.yaml
@@ -1,0 +1,4 @@
+- bump: minor
+  changes:
+    added:
+    - Create distinct Louisiana CDCC refundable and non-refundable variables.
