+++ policyengine_us/parameters/gov/states/la/tax/income/credits/cdcc/non_refundable/match.yaml
@@ -1,11 +1,10 @@
-description: Louisiana matches this percentage of the federal child and dependent care credit, depending on federal adjusted gross income.
+description: Louisiana matches this percentage of the federal child and dependent care credit, depending on federal adjusted gross income, as the non-refundable state CDCC.
 
 brackets:
   - threshold:
       2006-01-01: 0
     amount:
-      2006-01-01: 0.25
-      2007-01-01: 0.5
+      2006-01-01: 0
   - threshold:
       2006-01-01: 25_000
     amount:
@@ -19,14 +18,15 @@ brackets:
     amount:
       2006-01-01: 0.1
 
+
 metadata:
   type: single_amount
   threshold_unit: currency-USD
   amount_unit: /1
   period: year
-  label: Louisiana CDCC match
+  label: Louisiana CDCC non-refundable match
   reference:
-    - title: §297.4. Reduction to tax due; certain child care expenses, part (A)(2)-(4)
+    - title: §297.4. Reduction to tax due; certain child care expenses, part (A)(1)
       href: http://legis.la.gov/Legis/Law.aspx?d=101769
     - title: 2023 Louisiana Resident Income Tax Return, 2A
       href: https://revenue.louisiana.gov/TaxForms/IT540WEB(2023)%20F.pdf#page=16
