+++ policyengine_us/parameters/gov/states/la/tax/income/credits/cdcc/non_refundable/upper_bracket_cap.yaml
@@ -1,4 +1,4 @@
-description: Louisiana limits its child and dependent care credit to this maximum amount for filers in the credit's top income bracket.
+description: Louisiana limits its non-refundable child and dependent care credit to this maximum amount, for filers with income in the top bracket.
 
 values:
   2006-01-01: 25
@@ -6,7 +6,7 @@ values:
 metadata:
   period: year
   unit: currency-USD
-  label: Louisiana CDCC upper bracket max amount
+  label: Louisiana non-refundable CDCC upper bracket cap
   reference:
     - title: §297.4. Reduction to tax due; certain child care expenses, part (A)(4)
       href: http://legis.la.gov/Legis/Law.aspx?d=101769
