+++ policyengine_us/parameters/gov/states/la/tax/income/credits/non_refundable.yaml
@@ -1,6 +1,7 @@
 description: Louisiana provides the following non-refundable tax credits.
 values:
-  2021-01-01: []
+  2021-01-01: 
+    - la_non_refundable_cdcc
 metadata:
   unit: list
   period: year
