+++ policyengine_us/parameters/gov/states/la/tax/income/credits/refundable.yaml
@@ -1,7 +1,7 @@
 description: Louisiana provides the following refundable tax credits.
 values:
   2021-01-01:
-    - la_cdcc
+    - la_refundable_cdcc
     - la_eitc
 metadata:
   unit: list
