+++ policyengine_us/tests/policy/baseline/gov/states/la/tax/income/credits/cdcc/la_non_refundable_cdcc.yaml
@@ -1,0 +1,44 @@
+- name: AGI ineligible
+  period: 2006
+  input:
+    adjusted_gross_income: 25_000
+    cdcc: 1_000
+    state_code: LA
+  output:
+    la_non_refundable_cdcc: 0
+
+- name: 30% match above 25k
+  period: 2023
+  input:
+    adjusted_gross_income: 25_001
+    cdcc: 1_000
+    state_code: LA
+  output:
+    la_non_refundable_cdcc: 300
+
+- name: Match over 60k capped at $25
+  period: 2023
+  input:
+    adjusted_gross_income: 60_001
+    cdcc: 1_000
+    state_code: LA
+  output:
+    la_non_refundable_cdcc: 25
+
+- name: Match over 60k, uncapped
+  period: 2023
+  input:
+    adjusted_gross_income: 60_001
+    cdcc: 100
+    state_code: LA
+  output:
+    la_non_refundable_cdcc: 10
+
+- name: Match over 60k, uncapped
+  period: 2023
+  input:
+    adjusted_gross_income: 35_001
+    cdcc: 100
+    state_code: LA
+  output:
+    la_non_refundable_cdcc: 10
